Move note to seperate folder
