+++ lib/note/note.dart
